+++ examples/react/nextjs-suspense-streaming/src/app/page.tsx
@@ -1,11 +1,11 @@
 'use client'
-import { isServer, useSuspenseQuery } from '@tanstack/react-query'
+import { environmentManager, useSuspenseQuery } from '@tanstack/react-query'
 import { Suspense } from 'react'
 
 export const runtime = 'edge' // 'nodejs' (default) | 'edge'
 
 function getBaseURL() {
-  if (!isServer) {
+  if (!environmentManager.isServer()) {
     return ''
   }
   if (process.env.VERCEL_URL) {
@@ -30,7 +30,7 @@ function useWaitQuery(props: { wait: number }) {
     },
   })
 
-  return [query.data as string, query] as const
+  return [query.data, query] as const
 }
 
 function MyComponent(props: { wait: number }) {