+++ packages/angular-query-experimental/vite.config.ts
@@ -45,7 +45,7 @@ const config = defineConfig({
     environment: 'jsdom',
     setupFiles: ['test-setup.ts'],
     coverage: {
-      enabled: true,
+      enabled: !!process.env.CI,
       provider: 'istanbul',
       include: ['src/**/*'],
       exclude: ['src/__tests__/**'],