+++ packages/preact-query/src/types.ts
@@ -5,6 +5,7 @@ import type {
   DefinedQueryObserverResult,
   DistributiveOmit,
   FetchQueryOptions,
+  InfiniteData,
   InfiniteQueryObserverOptions,
   InfiniteQueryObserverResult,
   MutateFunction,
@@ -102,7 +103,7 @@ export type AnyUseInfiniteQueryOptions = UseInfiniteQueryOptions<
 export interface UseInfiniteQueryOptions<
   TQueryFnData = unknown,
   TError = DefaultError,
-  TData = TQueryFnData,
+  TData = InfiniteData<TQueryFnData>,
   TQueryKey extends QueryKey = QueryKey,
   TPageParam = unknown,
 > extends OmitKeyof<
@@ -127,7 +128,7 @@ export type AnyUseSuspenseInfiniteQueryOptions =
 export interface UseSuspenseInfiniteQueryOptions<
   TQueryFnData = unknown,
   TError = DefaultError,
-  TData = TQueryFnData,
+  TData = InfiniteData<TQueryFnData>,
   TQueryKey extends QueryKey = QueryKey,
   TPageParam = unknown,
 > extends OmitKeyof<