+++ packages/vue-query-devtools/vite.config.ts
@@ -23,7 +23,7 @@ const config = defineConfig({
     watch: false,
     environment: 'jsdom',
     coverage: {
-      enabled: true,
+      enabled: !!process.env.CI,
       provider: 'istanbul',
       include: ['src/**/*'],
       exclude: ['src/__tests__/**'],